requirements.txt
fastapi
uvicorn
websockets
loguru
transformers
pydantic
accelerate>=0.26.0
python-multipart
pydub
pyannote.audio
rich